pyproject.toml
[tool.poetry]
name = "warp-plus-bot"
version = "0.1.0"
description = "Telegram bot which can give you free WARP+ bandwidth."
authors = ["Culture Cloud"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
APScheduler = "^3.7.0"
Pyrogram = "^1.2.9"
python-dotenv = "^0.18.0"
requests = "^2.26.0"
TgCrypto = "^1.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
